+++ src/sink/file.rs
@@ -72,7 +72,7 @@ impl Sink for File {
     async fn on_event(
         &mut self,
         _input: &str,
-        codec: &dyn Codec,
+        codec: &mut dyn Codec,
         _codec_map: &HashMap<String, Box<dyn Codec>>,
         mut event: Event,
     ) -> ResultVec {